Add function for explicit virtual operation access check
Currently it is easy for a virtual command to check access of the
bound principal to the associated virtual operation, but there is no
nice way to evaluate access of an arbitrary principal to a
particular virtual operation.  We will need something like this for
Dogtag-based certificate request validation.

Add the 'check_operation_access' method for explicit virtual
operation access checks.  Refactor 'VirtualCommand.check_access()'
to use it internally.

Part of: https://pagure.io/freeipa/issue/5011
Part of: https://pagure.io/freeipa/issue/6423