cert-request: convey subject principal to Dogtag
When we switch to GSS-API authentication to Dogtag, the subject
principal name will need to be explicitly conveyed to Dogtag as
arbitrary user data (which will then be passed on to the
ipa-pki-validate-cert-request program via the
ExternalProcessConstraint).

Convey the subject principal to Dogtag via the 'user-data' query
param.

Part of: https://pagure.io/freeipa/issue/5011