Deserialise Python exceptions in request_certificate response
When we switch to GSS-API authentication to Dogtag, certificate
request authorisation and validation will be performed by Dogtag
invoking the ipa-pki-validate-cert-request program.  This program
outputs a serialised PublicError on failure, which gets reflected in
the response from Dogtag.

Detect if a failed certificate request contains a serialised
PublicError in the response, and if it does, construct and raise it.

Part of: https://pagure.io/freeipa/issue/5011