cert-request: validate SRVName in CSR
The SRVName otherName type is provides an alternative to dNSName as
a means of certificate subject validation.  It can be used when a
client has located the service via a DNS SRV record.[1]

Some protocols require or recommend the use of SRVName in
certificates.  These include XMPP[2], TZDIST[3] and email-related
protocols where SRV-based service discovery is supported[4].

[1] https://tools.ietf.org/html/rfc6125
[2] https://tools.ietf.org/html/rfc6120#section-13.7.1.2
[3] https://tools.ietf.org/html/rfc7808#section-8
[4] https://tools.ietf.org/html/rfc7817#section-5

Therefore it is desirable for FreeIPA to support issuing
certificates with SRVName.  SRVName values in a CSR are checked
against all principal aliases of the subject principal.  Aliases
that are non-host service principals are candidates, and the service
name and DNS name parts are compared against the service principals
service and hostname parts respectively.  The comparison is case
insensitive.

The Kerberos realm is ignored.  Unlike DNSName validation, no
attempt is made to locate an "alternative principal" for names that
do not match the nominated subject principal.  (This heuristic
predates Kerberos principal alias support in FreeIPA and is kept for
backwards compatibility only).